Move profile editor data retrieval code into UserProfileService.
